pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["ph <robotnaoborot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.1.5"
django-rest-framework = "^0.1.0"
django_filter = "^2.4.0"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
